Refactored to use a function for looking up value for a label
The reason for this switch is that special case logic is needed to determine the US state name to display.  All other fields can be looked up by name.

The burden of fallback functionality is now in `helper/labelSchema` so the burden of unit testing is now in `test/unit/helper/labelSchema.js`