Handle fatal errors in unit tests
Due to the way our unit tests are run, a fatal error in the API code
that causes the node process to quit with an error will not be picked up
by `tap-dot`, and the unit tests will still return a passing status
code. This is due to the way status codes from pipes are handled by
default.

Fortunately, there is an ["unofficial bash strict mode"](http://redsymbol.net/articles/unofficial-bash-strict-mode/)
that helps fix this sort of thing, and by running our unit tests with a
few tweaks we can ensure that any failure in the unit tests is captured.