Fix timeout error handling
Timeout errors come back from the Elasticsearch module as Error objects,
so instead of pushing the entire object onto the errors array, we should
just push the message.

I'm not sure if there are other cases where errors are just strings,
errors that aren't objects with a message property are handled as
before.