Configurable contribution of name in confidence scoring
The influence of name property is enhanced in confidence scoring,
because it is an essential part of the search. Now the name is
normalized lightly before checking the match (e.g. to fix street/number
order issues), and the contribution range starts from zero instead
of 0.7. Upper limit is defined by:

  pelias_config.api.localization.confidenceNameWeight

Default value is 1 as before, but a higher value such as 2 seems
to cure some strange ordering problems originating from scoring
peculiarities of ES multi_match searches.