Use lower confidence for deal breaker filtered entries
Confidence for entries failing the deal breaker test was 0.5.
This is a high value. Entries, which pass the test may get much
lower confidence. E.g. search 'posti, 33100' in Finland gives
confidence of ~ 0.3 for quite close matches in the mentioned zip
code area. Now score 0.1 is assigned to the failed entries.