Only call language service when language not defaulted
The default language is English, which is also what is stored in
Elasticsearch. Thus, if the user did not specify a language via headers
or query parameter, there is no reason to call Placeholder.

Some Placeholder responses for language translation can be 30KB, and all
that JSON takes considerable time to parse