Remove focus_selected_layers query view
This query extends the standard focus query view with hardcoded layers
for which the query applies. The intent was to apply the focus scoring
only to non-admin areas, but the list of layers was already out of date,
as it was missing streets.

The query is fundamentally problematic with custom layers as well.