fix(predicate): Fix address related layers detection
This fixes an error in the logic from
https://github.com/pelias/api/pull/1307.

The `address_search_with_ids` query can only return results in the
address or street layers. However, the predicate logic was allowing the
query to execute if _any_ admin layer was specified, which could allow
such a query to return street or address layers when not intended.