feat(addressSearchUsingIds): Set default street slop to 1
Combined with https://github.com/pelias/query/pull/101, this change
allows for improved matching of streets when using the
addressSearchUsingIDs query (the first query executed by the search
endpoint).

Essentially, by allowing a `slop` of 1 in the `match_phrase` clauses of
the Elasticsearch query, its now possible to omit a single word _in the
middle_ of a street name and still match it. Omitting words at the start
or end of the street name has been supported since
https://github.com/pelias/schema/pull/310.

Some examples of where this helps:

| Actual street name | Newly matching input text |
| ---                | ---                       |
| SE Martin Luther King Jr Blvd | Martin Luther King Blvd |
| Carrer de Balmes | Carrer Balmes |