feat(place): Query for old and new contents of `_id` field
_This change completes step 2 of the `_id` field contents migration
described in https://github.com/pelias/pelias/issues/672#issuecomment-508837195 _

As part of the transition away from Elasticsearch types, we must change
the contents of the `_id` field in our Elasticsearch documents.

This change modifies the `/v1/place` endpoint so that it queries
Elasticsearch for documents based on both the old and new contents of
the `_id` field, allowing the same API code to work regardless of the
version of Pelias used to import data.

Connects https://github.com/pelias/pelias/issues/672