feat(GID): Require GID to be stored in the `_id` field
This change implements the 5th and final step in the GID migration plan
outlined in
https://github.com/pelias/pelias/issues/672#issuecomment-508837195.

It adjusts the expectations of the API so that it requires the
full GID, rather than the source_id, to be stored in the `_id` field in
Elasticsearch.

In the process, the `source_id` field is also no longer used in
`helper/geojsonify.js`, making this the first step towards removing the
`source_id` field (https://github.com/pelias/schema/issues/383).

Connects https://github.com/pelias/pelias/issues/672