feat(query): Use match_all for subquery matching
This PR brings the `search_pelias_parser` queries into alignment with
autocomplete and `search` queries by using the `match_all` query,
instead of a phrase query, for use with function scoring to apply boosts
for popularity and population.

While in theory this could result in longer query times because the
population and popularity scoring will be applied to more documents, its
unlikely we'll notice that.

However, what we _would_ notice is if there's highly populated or
popular record that does not match the phrase subquery, but does match
the non-phrase query.