feat(place): Only query for new GID based _ids
BREAKING_CHANGE: Importers using pelias-model 6.0.0 or greater
(https://github.com/pelias/model/pull/121) are required with this
change.

_This change completes step 4 of the Elasticsearch _id migration plan
laid out in
https://github.com/pelias/pelias/issues/672#issuecomment-508837195_

This is a follow up to https://github.com/pelias/api/pull/1329 which
removes the querying of the "old" `_id` contents, requiring the new
style, where the full Pelias GID is stored in the Elasticsearch ID
field.