feat(breaking): Retroactive version bump for breaking change
BREAKING CHANGE: A breaking change was made in d1b3976b as part of
https://github.com/pelias/api/pull/1363, but was missed in the release
notes and version increment. This release is meant to draw attention to
that breaking change.

Any Pelias users looking to use pelias/api version 3.86.0 or later will
need to ensure they use the API in combination with an Elasticsearch
build run using compatible imports.

As long as the importers are using a version released on September 13th,
2019 or later, there will be no compatibility issues.

Here is the full list of compatible importer version numbers:
* openaddresses: 2.37.0
* openstreetmap: 4.29.0
* whosonfirst: 2.44.1
* geonames: 2.32.2
* polylines: 1.12.2
* csv: 1.7.3

For full details on the breaking change, see the following issues and
pull requests:

* https://github.com/pelias/pelias/issues/672#issuecomment-508837195
* https://github.com/pelias/api/pull/1363
* https://github.com/pelias/api/pull/1364

The following issue track the errors Pelias users will see if they are
using incompatible versions of the API and importers:
https://github.com/pelias/api/issues/1379.

Our apologies for the confusion caused by anyone hit by this!