fix(log): Correctly record Elasticsearch result count
This was essentially logging how many results had already been set on the API
response under construction. So it was sometimes returning a number, but
it was actually logging based on the results of previous invocations of
the search controller. This isn't super common, so the number was
usually zero.

Fixes https://github.com/pelias/api/issues/1386