Fall back to pelias parser only when it may have better results
The Pelias Parser and the query that uses it may sometimes have better
results than the earlier queries. However, this is only true in specific
circumstances, mostly intersection queries.

This change has the `defer_to_pelias_parser` attempt to detect cases in
which the Pelias Parser _won't_ do a better job, and avoid using it in
those cases.