add a "restrictIds" function to pelias for debugging
The purpose of this is that often I want to know "why did document A
score above document B" - pulling open the entire ES debugging response
(especially when nested in the pelias response) makes it super hard to
keep track of what document I'm looking at.

With this change, internal links are added that ask pelias to run the
same query, just restricted to that document. Such that two of these can
be opened side by side for relatively easy comparison of scoring.