feat(focus.point): ensure scoring is applied to all results
We create queries to handle the `focus.point` parameter by using the
[function_score](https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-function-score-query.html)
elasticsearch query. `function_score` requires a function, and a query.

The original structure of these queries was written long ago, when we
didn't understand Elasticsearch nearly as well. They use a
`match_phrase` query based on some of the input text for this query.
However, there's really no reason to do so, we want the function score
for distance from the focus point to apply to _all_ results, not just
the subset of results that match the `match_phrase` query.

Usually, the subset of results that has the focus point scoring applied
is almost 100% of all results, but not always, which can be very
confusing.