add static serving of files in frontend/
The idea of this change is to make it easy for developers to add their own custom on-server debug pages to pelias. A major use cases for this in my opinon is to be able to include a page with server build information and another with index debug information. A developer can generate pages and stick them in frontend as part of their build/deploy process without needing to get their changes integrated into mainlien pelias.

We discusses this change as a replacement for an earlier PR that included server build info in the build process.