feat(autocomplete): Add support for autocompleting postalcodes
Enable support for autocompleting postalcode by relaxing constraints
around when we allow partial matches on numeric inputs.

Historically, Pelias has not allowed the `text` parameter input to match
a "partial" result when the entire input consists only of numbers.
Considering there are often _lots_ of results that could match a numeric
input like `1`, `10`, or even `1014`, and most of them are addresses
that have little chance of being relevant, this approach isn't without
merit.

However postalcodes are another possible match, and people want to
search for them reasonably often. They also often consist purely of
numbers.

So this change allows those numeric partial matches when the `address`
layer _won't_ be queried. Thanks to some of our performance
optimizations, this will always be the case for short inputs like
`text=1234` as long as the `layers` parameter is unset or doesn't
explicitly contain the address layer.

In practice, what this means is that a query like `text=9021` will now
result in the `90210` US postalcode first, whereas previously it would
return a bunch of irrelevant venue results scattered around the world.