fix(config): Remove default timeout and retry values
The `pelias.json` schema validation in the API has been setting default
values for the `timeout` and `retries` parameters used for network calls
to each of the other Pelias services.

Years ago, in https://github.com/pelias/microservice-wrapper/pull/30, we
determined that the original defaults of 250ms were too low.

However, the schema defaults were still using this 250ms value, which
means they were effectively overriding the new, more sensible default of
1000ms in the microservice-wrapper library.

This change removes the defaults for both the `timeout` and `retries`
values from the Pelias API, meaning the defaults set in the
`microservice-wrapper` itself will be used.