Remove Mac OS from CI build matrix
This was a nice idea, but testing on a Mac is not especially relevant
for Node.js code, so there's not much benefit.

While build minutes for _open source_ projects on GitHub are free,
anyone wanting to run these tests in a _private_ repo probably won't
want to test on a Mac either, as Mac build minutes cost **10x** more
than Linux!