fix(query): ignore empty tokens
Empty tokens should not make it this far into the query code, but it can
happen.

In that case, we should ignore empty tokens and avoid aggrivating the
`VariableStore` checks by attempting to store empty strings.

This fixes the 500 erro described in
https://github.com/pelias/api/issues/1535, but doesn't really fix the
underlying problem.

From my investigation it looks like Emoji that include tokens like a
[variation selector](https://emojipedia.org/variation-selector-16/) will
not be completely removed by the Pelias Parser (the emoji codepoint will
be removed, but not the variation selector).