Run Docker image build _after_ release
As discussed in https://github.com/pelias/ci-tools/pull/7, we need to
build the Docker images _after_ we publish any new versions, so that the
git tag for the version is present.

This lets us publish a Docker image corresponding to the tag.

The `if` check will ensure the unit tests passed, and that the
npm-publish job either succeeded (as when running on the master branch),
or was skipped (non-master branches).

Note that the call to the `always()` function is required, otherwise
GitHub applies a default check that we don't want before deciding
whether or not to run a job. See the somewhat cryptic mention of it in
the docs here:
https://docs.github.com/en/actions/reference/context-and-expression-syntax-for-github-actions#job-status-check-functions