fix(services): Return 502 response code for service timeouts
Pelias has always had a bit of trouble selecting the right HTTP response
code in the face of various error states.

Up until https://github.com/pelias/api/pull/1231 in 2018, we reported
almost all timeouts from slow Elasticsearch queries as HTTP 400 errors,
not something in the more appropriate 5XX range. This suggests to
consumers of the Pelias API that they made a mistake in calling Pelias,
instead of the reality that Pelias was just being slow.

Even after that change, it turns out we were _still_ classifying
timeouts to other Pelias services (like Placeholder or Interpolation) as
400 errors instead of 5XX. All the Pelias services are generally very
fast, so this was not nearly as much of an issue, but timeouts do
happen.

This PR adds additional handling to detect timeout errors and give them
their own subclass of `Error` that can be treated appropriately
everywhere. Timeouts waiting for any Pelias service will now return HTTP
502 errors just like a timeout waiting for Elasticsearch.