Pin to an exact pelias-parser version (2.1.0)
Historically we've kept most Pelias NPM dependencies on carat versions,
but because the Pelias Parser can cause changes in acceptance-test
results for any change (minor or even bug fix), we've kept it pinned to
a specific version so we can explicitly upgrade.

Recently, we were on a [hotfix
branch](https://github.com/pelias/api/pull/1531) of the parser, but
while switching back to using a version specification, we switched to
[using a carat
range](https://github.com/pelias/api/pull/1565).

While manually bumping the version on a frequent basis is a bit
annoying, it helps avoid any chance that unexpected or untested parser
changes will find their way into the API releases, so this change brings
us back to a pin of the current latest version of the Pelias Parser.