fix(sanitizer): Account for new parameter in categories sanitizer
It turns out the changes to the parameters for sanitizer functions in
https://github.com/pelias/api/pull/1583 wasn't _quite_ backwards
compatible.

The `categories` parameter sanitier supports taking a 3rd parameter that
defines how to determine valid categories. If there's only two
parameters, it uses a default function that always returns true.

We intentionally haven't documented the categories parameter because
it's not really done, and the categories mapping is not finalized. As a
result almost no one uses it.

However, if they did use it, and passed in a category, they could cause
the API to generate a 500 error.