feat(place): Support any casing for GID sources and layers
We have always supported any combination of uppercase or lowercase for
the `sources` and `layers` parameters.

However, the `ids` parameter to the `/v1/place` endpoint has required
that the GID be written in lowercase.

While it seems odd that anyone would ever send Pelias a GID in any
format except what Pelias generates, apparently this does happen.

Lowercasing the source and layer components (but not the ID part) of a
GID seems like a nice thing to do, so we might as well.