feat(perf): Use only phrase.default field for search
This avoids use of the expensive `name.default` field, which includes
partial tokens (ngrams)

This change is extremely beneficial to performance of search queries,
especially those that are "gibberish".

In particular, it avoids the huge number of document matches that come
from searching for addresses with the housenumber `1`. With the
`name.default` field in use, all records with a token that _starts with_
1 would be returned. That's generally not what we want for the search
  endpoint.