feat(dedupe): Do not deduplicate US States
It's common for US states to have either a county or city within them
that shares a name (with minor possible differences).

For example:
Arksansas City in Arkansas
Hawaii County in Hawaii
Iowa City in Iowa
California City in California
New York City in New York (of course!)

Our general deduplication logic considers an admin record that is
parented by a record sharing its name to be the same. This works well
for places like Singapore, Berlin, and Tokyo, which all have a city
(locality) that is conceptually the same as a region or country.

US states, however, are _not_ conceptually the same as any of these
cities, and they should pretty much always show up in results.

This PR adds a check that the record is _not_ a US state before
performing the general heierarchy checks. There's one exception: US
states can dedupe against other US states, so that Geonams and WOF
records can deduplicate themselves.

This PR allows us to merge https://github.com/pelias/api/pull/1371