feat(cluster): Allow multiple worker processes with cluster module
This adds the ability for the API to use multiple worker processes with
Node.js's builtin [`cluster`](https://nodejs.org/api/cluster.html)
module.

Historically, I've been opposed to adding this functionality, preferring
that Pelias users manage parallelism on their own such as with
Kubernetes or something else.

But there are enough use cases where that sort of orchestration isn't
worth the complexity, and you want more than one CPU's worth of API
requests.

This implementation is essentially lifted from the one in Placeholder,
with mostly cosmetic changes only.

The biggest difference is the default. For backwards compatibility,
without specifing the new `CPUS` environment variable, the API will
_not_ use the cluster module and operate as a single process just like
before.

_With_ the `CPUS` variable set to a number, that many worker processes
will be launched, up to the number of CPUs detected on the machine.