Use GitHub Actions variable for Ubuntu version
This is a new approach to managing our Ubuntu version for CI where the
version comes from a Github Actions variable instead of being hardcoded.

Hopefully this makes it easier for us to change the version in one go
across all our repositories instead of having to make dozens of commits
that are similar.

The variable's value is currently set to ubuntu-22.04, so this also
handles https://github.com/pelias/pelias/issues/951

https://github.com/organizations/pelias/settings/variables/actions