fix(layers): Handle sources and layers param with address filter
As reported in https://github.com/pelias/api/issues/1653, there is an
issue when negative sources are specified in combination with negative
layers. The end result is that the negative layers are ignored.

This issue was likely introduced in
https://github.com/pelias/api/pull/1604 or
https://github.com/pelias/api/pull/1525, the pull requests where we
added the concept of negative sources and layers and subsequently
improved them.

Fixes https://github.com/pelias/api/pull/1525