Skip language service lookup for parent IDs same as the record
WOF and Geonames andmin records usually have a parent record in the same
layer as their own which lists their own ID.

For example the Geonames record for NYC has its own ID in the
locality_id field.

If present, any translations for the record itself will already be on
the record, so there is no sense in looking them up.

It especially makes no sense for Geonames records, since the language
service does not have any translations for Geonames at all.